+++ src/Elcodi/Admin/GeoBundle/Controller/CountryController.php
@@ -78,6 +78,10 @@ class CountryController extends AbstractAdminController
         $orderByField,
         $orderByDirection
     ) {
+        if (!$this->canRead()) {
+            throw $this->createAccessDeniedException();
+        }
+
         return [
             'page'             => $page,
             'limit'            => $limit,
@@ -153,6 +157,16 @@ class CountryController extends AbstractAdminController
         $isValid,
         Request $request
     ) {
+        if ($country->id) {
+            if (!$this->canUpdate()) {
+                throw $this->createAccessDeniedException();
+            }
+        } else {
+            if (!$this->canCreate()) {
+                throw $this->createAccessDeniedException();
+            }
+        }
+        
         if ($isValid) {
 
             $this->flush($country);
@@ -205,6 +219,10 @@ class CountryController extends AbstractAdminController
         $entity,
         $redirectPath = null
     ) {
+        if (!$this->canDelete()) {
+            throw $this->createAccessDeniedException();
+        }
+
         return parent::deleteAction(
             $request,
             $entity,
