+++ src/Elcodi/Admin/ProductBundle/Controller/CategoryController.php
@@ -38,8 +38,6 @@ use Elcodi\Component\Product\Entity\Interfaces\CategoryInterface;
  */
 class CategoryController extends AbstractAdminController
 {
-    private $resource = "category";
-
     /**
      * List elements of certain entity type.
      *
@@ -57,7 +55,7 @@ class CategoryController extends AbstractAdminController
      */
     public function listAction()
     {
-        if (!$this->canRead($this->resource)) {
+        if (!$this->canRead()) {
             throw $this->createAccessDeniedException();
         }
         
@@ -131,6 +129,17 @@ class CategoryController extends AbstractAdminController
         $isValid,
         Request $request
     ) {
+        if ($category->id) {
+            if (!$this->canUpdate()) {
+                throw $this->createAccessDeniedException();
+            }
+        } else {
+            if (!$this->canCreate()) {
+                throw $this->createAccessDeniedException();
+            }
+        }
+        
+
         if ($isValid) {
             $this->flush($category);
 
@@ -179,6 +188,10 @@ class CategoryController extends AbstractAdminController
         Request $request,
         EnabledInterface $entity
     ) {
+        if (!$this->canUpdate()) {
+            throw $this->createAccessDeniedException();
+        }
+
         return parent::enableAction(
             $request,
             $entity
@@ -210,6 +223,10 @@ class CategoryController extends AbstractAdminController
         Request $request,
         EnabledInterface $entity
     ) {
+        if (!$this->canUpdate()) {
+            throw $this->createAccessDeniedException();
+        }
+
         return parent::disableAction(
             $request,
             $entity
@@ -243,6 +260,10 @@ class CategoryController extends AbstractAdminController
         $entity,
         $redirectPath = null
     ) {
+        if (!$this->canDelete()) {
+            throw $this->createAccessDeniedException();
+        }
+
         return parent::deleteAction(
             $request,
             $entity,
