+++ src/Elcodi/Admin/ProductBundle/Controller/ManufacturerController.php
@@ -78,6 +78,10 @@ class ManufacturerController extends AbstractAdminController
         $orderByField,
         $orderByDirection
     ) {
+        if (!$this->canRead()) {
+            throw $this->createAccessDeniedException();
+        }
+
         return [
             'page'             => $page,
             'limit'            => $limit,
@@ -153,6 +157,17 @@ class ManufacturerController extends AbstractAdminController
         $isValid,
         Request $request
     ) {
+        if ($manufacturer->id) {
+            if (!$this->canUpdate()) {
+                throw $this->createAccessDeniedException();
+            }
+        } else {
+            if (!$this->canCreate()) {
+                throw $this->createAccessDeniedException();
+            }
+        }
+        
+
         if ($isValid) {
             $firstImage = $manufacturer
                 ->getSortedImages()
@@ -210,6 +225,10 @@ class ManufacturerController extends AbstractAdminController
         Request $request,
         EnabledInterface $entity
     ) {
+        if (!$this->canUpdate()) {
+            throw $this->createAccessDeniedException();
+        }
+
         return parent::enableAction(
             $request,
             $entity
@@ -241,6 +260,10 @@ class ManufacturerController extends AbstractAdminController
         Request $request,
         EnabledInterface $entity
     ) {
+        if (!$this->canUpdate()) {
+            throw $this->createAccessDeniedException();
+        }
+
         return parent::disableAction(
             $request,
             $entity
@@ -274,6 +297,10 @@ class ManufacturerController extends AbstractAdminController
         $entity,
         $redirectPath = null
     ) {
+        if (!$this->canDelete()) {
+            throw $this->createAccessDeniedException();
+        }
+
         return parent::deleteAction(
             $request,
             $entity,
